layer: Move some debug report defs/funcs to header file
