+++ layer/profiles.cpp
@@ -186,80 +186,6 @@ static std::string GetSimulateCapabilitiesLog(SimulateCapabilityFlags flags) {
     return result;
 }
 
-enum DebugAction {
-    DEBUG_ACTION_FILE_BIT = (1 << 0),
-    DEBUG_ACTION_STDOUT_BIT = (1 << 1),
-    DEBUG_ACTION_OUTPUT_BIT = (1 << 2),
-    DEBUG_ACTION_BREAKPOINT_BIT = (1 << 3)
-};
-typedef int DebugActionFlags;
-
-static DebugActionFlags GetDebugActionFlags(const vku::Strings &values) {
-    DebugActionFlags result = 0;
-
-    for (std::size_t i = 0, n = values.size(); i < n; ++i) {
-        if (values[i] == "DEBUG_ACTION_FILE_BIT") {
-            result |= DEBUG_ACTION_FILE_BIT;
-        } else if (values[i] == "DEBUG_ACTION_STDOUT_BIT") {
-            result |= DEBUG_ACTION_STDOUT_BIT;
-        } else if (values[i] == "DEBUG_ACTION_OUTPUT_BIT") {
-            result |= DEBUG_ACTION_OUTPUT_BIT;
-        } else if (values[i] == "DEBUG_ACTION_BREAKPOINT_BIT") {
-            result |= DEBUG_ACTION_BREAKPOINT_BIT;
-        }
-    }
-
-    return result;
-}
-
-static std::string GetDebugActionsLog(DebugActionFlags flags) {
-    std::string result = {};
-
-    if (flags & DEBUG_ACTION_FILE_BIT) {
-        result += "DEBUG_ACTION_FILE_BIT";
-    }
-    if (flags & DEBUG_ACTION_STDOUT_BIT) {
-        if (!result.empty()) result += ", ";
-        result += "DEBUG_ACTION_STDOUT_BIT";
-    }
-    if (flags & DEBUG_ACTION_OUTPUT_BIT) {
-        if (!result.empty()) result += ", ";
-        result += "DEBUG_ACTION_OUTPUT_BIT";
-    }
-    if (flags & DEBUG_ACTION_BREAKPOINT_BIT) {
-        if (!result.empty()) result += ", ";
-        result += "DEBUG_ACTION_BREAKPOINT_BIT";
-    }
-
-    return result;
-}
-
-enum DebugReport {
-    DEBUG_REPORT_NOTIFICATION_BIT = (1 << 0),
-    DEBUG_REPORT_WARNING_BIT = (1 << 1),
-    DEBUG_REPORT_ERROR_BIT = (1 << 2),
-    DEBUG_REPORT_DEBUG_BIT = (1 << 3)
-};
-typedef int DebugReportFlags;
-
-DebugReportFlags GetDebugReportFlags(const vku::Strings &values) {
-    DebugReportFlags result = 0;
-
-    for (std::size_t i = 0, n = values.size(); i < n; ++i) {
-        if (values[i] == "DEBUG_REPORT_NOTIFICATION_BIT") {
-            result |= DEBUG_REPORT_NOTIFICATION_BIT;
-        } else if (values[i] == "DEBUG_REPORT_WARNING_BIT") {
-            result |= DEBUG_REPORT_WARNING_BIT;
-        } else if (values[i] == "DEBUG_REPORT_ERROR_BIT") {
-            result |= DEBUG_REPORT_ERROR_BIT;
-        } else if (values[i] == "DEBUG_REPORT_DEBUG_BIT") {
-            result |= DEBUG_REPORT_DEBUG_BIT;
-        }
-    }
-
-    return result;
-}
-
 #define APPEND(name)                         \
     if (flags & name) {                      \
         if (!result.empty()) result += ", "; \
@@ -6967,16 +6893,16 @@ static void InitSettings(const void *pnext) {
     layer_settings.exclude_device_extensions.clear();
     layer_settings.exclude_formats.clear();
 
-    const VkProfileLayerSettingsEXT *settings;
+    const VkProfileLayerSettingsEXT *user_settings;
     // Programmatically specified settings override ENV vars or layer settings file settings
-    if ((pnext) && (settings = FindSettingsInChain(pnext))) {
-        layer_settings.profile_file = settings->profile_file;
-        layer_settings.profile_name = settings->profile_name;
-        layer_settings.profile_validation = settings->profile_validation;
-        layer_settings.simulate_capabilities = settings->simulate_capabilities;
-        layer_settings.debug_fail_on_error = settings->debug_fail_on_error;
-        layer_settings.exclude_device_extensions = settings->exclude_device_extensions;
-        layer_settings.exclude_formats = settings->exclude_formats;
+    if ((pnext) && (user_settings = FindSettingsInChain(pnext))) {
+        layer_settings.profile_file = user_settings->profile_file;
+        layer_settings.profile_name = user_settings->profile_name;
+        layer_settings.profile_validation = user_settings->profile_validation;
+        layer_settings.simulate_capabilities = user_settings->simulate_capabilities;
+        layer_settings.debug_fail_on_error = user_settings->debug_fail_on_error;
+        layer_settings.exclude_device_extensions = user_settings->exclude_device_extensions;
+        layer_settings.exclude_formats = user_settings->exclude_formats;
     } else {
 
         if (vku::IsLayerSetting(kOurLayerName, kLayerSettingsProfileFile)) {
